# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Tech Blog
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to the weekly newsletter (sends confirm email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscribeResponse"
...
  /admin/subscribers/{id}:
    delete:
      operationId: deleteSubscriber
      tags: [newsletter]
      summary: Delete a subscriber (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/newsletter/test:
    post:
      operationId: sendTestNewsletter
      tags: [newsletter]
      summary: Send a test digest to a single email (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterTestBody"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterActionResponse"
  /admin/newsletter/send-now:
    post:
      operationId: sendNewsletterNow
      tags: [newsletter]
      summary: Send the editor-composed digest immediately (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSendNowBody"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterActionResponse"
...
components:
  schemas:
...
    NewsletterSubscribeBody:
      type: object
      properties:
        email:
          type: string
        source:
          type: string
      required:
        - email
    NewsletterSubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    NewsletterTestBody:
      type: object
      properties:
        to:
          type: string
        subject:
          type: string
        editorNote:
          type: string
      required:
        - to
        - subject
    NewsletterSendNowBody:
      type: object
      properties:
        subject:
          type: string
        editorNote:
          type: string
      required:
        - subject
    NewsletterActionResponse:
      type: object
      properties:
        success:
          type: boolean
        sent:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
        posts:
          type: integer
        postCount:
          type: integer
        message:
          type: string
      required:
        - success
...